feature: do not load zaakafhandel details
Reduce zaakafhandel cache memory footprint
Introduce ZaakafhandelParametersSummary entity

Solves: PZ-3034
