+++ src/main/kotlin/net/atos/zac/smartdocuments/templates/model/SmartDocumentsTemplate.kt
@@ -14,7 +14,7 @@ import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 import jakarta.persistence.SequenceGenerator
 import jakarta.persistence.Table
-import net.atos.zac.admin.model.ZaakafhandelParameters
+import net.atos.zac.admin.model.ZaakafhandelParametersSummary
 import net.atos.zac.util.FlywayIntegrator
 import nl.lifely.zac.util.AllOpen
 import java.time.ZonedDateTime
@@ -50,7 +50,7 @@ class SmartDocumentsTemplate {
 
     @ManyToOne
     @JoinColumn(name = "zaakafhandelparameters_id", nullable = false)
-    lateinit var zaakafhandelParameters: ZaakafhandelParameters
+    lateinit var zaakafhandelParameters: ZaakafhandelParametersSummary
 
     @Column(name = "informatie_object_type_uuid", nullable = false)
     lateinit var informatieObjectTypeUUID: UUID
